pyproject.toml
[tool.poetry]
name = "instagram-reel-builder"
version = "0.1.0"
description = "Create Instagram reels from images with automatic 9:16 cropping and transitions"
authors = ["sfelaco"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11"
gradio = ">=4.0.0"
pillow = ">=10.0.0"
opencv-python = ">=4.8.0"
numpy = ">=1.24.0"
moviepy = ">=1.0.3"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
